Incorporate fs copy
