+++ crates/init/src/copy.rs
@@ -1,6 +1,8 @@
-use anyhow::{bail, Context, Result};
-use std::{ffi::OsStr, path::Path};
-use tokio::process::Command;
+use anyhow::Result;
+use e3_fs::prelude::*;
+use e3_fs::Fs;
+use regex::Regex;
+use std::path::Path;
 
 #[derive(Debug, Clone)]
 pub struct Filter {
@@ -8,7 +10,6 @@ pub struct Filter {
     pub search_pattern: String,
     pub replacement: String,
 }
-
 impl Filter {
     pub fn new(glob_pattern: &str, search_pattern: &str, replacement: &str) -> Self {
         Filter {
@@ -19,132 +20,185 @@ impl Filter {
     }
 }
 
+pub async fn copy_with_filters_impl<P1, P2>(
+    fs: &Fs,
+    src_path: P1,
+    dest_path: P2,
+    filters: &[Filter],
+) -> Result<()>
+where
+    P1: AsRef<Path> + Send + Sync,
+    P2: AsRef<Path> + Send + Sync,
+{
+    fs.mkdirp(&dest_path).await?;
+    fs.cp(&src_path, &dest_path).await?;
+    for filter in filters {
+        let file_paths = fs.find_files(&dest_path, &filter.glob_pattern).await?;
+        for file_path in file_paths {
+            fs.replace_in_place(
+                &Regex::new(&filter.search_pattern)?,
+                &filter.replacement,
+                file_path,
+            )
+            .await?;
+        }
+    }
+    Ok(())
+}
+
 pub async fn copy_with_filters<P1, P2>(
     src_path: P1,
     dest_path: P2,
     filters: &[Filter],
 ) -> Result<()>
 where
-    P1: AsRef<Path>,
-    P2: AsRef<Path>,
+    P1: AsRef<Path> + Send + Sync,
+    P2: AsRef<Path> + Send + Sync,
 {
-    let src_path = src_path.as_ref();
-    let dest_path = dest_path.as_ref();
-
-    // Create destination directory and copy contents in one shell command
-    let combined_command = format!(
-        "mkdir -p {} && cp -r {}/. {}",
-        dest_path.to_string_lossy(),
-        src_path.to_string_lossy(),
-        dest_path.to_string_lossy()
-    );
-
-    let output = Command::new("sh")
-        .arg("-c")
-        .arg(&combined_command)
-        .output()
-        .await
-        .context("Failed to execute mkdir and cp commands")?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        bail!("mkdir and cp commands failed: {}", stderr);
-    }
-
-    // Then apply filters to matching files
-    for filter in filters {
-        apply_filter_to_files(dest_path.as_os_str(), filter).await?;
-    }
+    copy_with_filters_impl(&Fs::physical()?, src_path, dest_path, filters).await?;
     Ok(())
 }
 
-// Async function to detect if we're using BSD sed
-// You can test for BSD sed by checking if the --version flag is supported.
-// GNU sed supports --version while BSD sed doesn't and will exit with an error.
-async fn is_bsd_sed() -> bool {
-    Command::new("sed")
-        .arg("--version")
-        .output()
-        .await
-        .map(|output| !output.status.success())
-        .unwrap_or(false)
-}
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use anyhow::Result;
+    use e3_fs::Fs;
+    #[tokio::test]
+    async fn test_copy_with_filters_impl() -> Result<()> {
+        // Create an in-memory filesystem
+        let fs = Fs::mem();
+        // Define source and destination paths
+        let src_path = "/tmp/mysource/.";
+        let dest_path = "/home/user/myproj";
+        fs.mkdirp(&dest_path).await?;
+        // Create some test files in the source directory
+        fs.mkdirp(&src_path).await?;
+        fs.write_to_file(format!("{}/file1.txt", src_path), "Content of file 1")
+            .await?;
+        fs.write_to_file(format!("{}/file2.txt", src_path), "Content of file 2")
+            .await?;
+        fs.mkdirp(format!("{}/subdir", src_path)).await?;
+        fs.write_to_file(
+            format!("{}/subdir/file3.txt", src_path),
+            "Content of file 3",
+        )
+        .await?;
+        // Create a nested directory structure with a package.json file
+        fs.mkdirp(format!(
+            "{}/tools/build/scripts/utils/helper-tool",
+            src_path
+        ))
+        .await?;
+        fs.write_to_file(
+            format!(
+                "{}/tools/build/scripts/utils/helper-tool/package.json",
+                src_path
+            ),
+            r#"{
+  "name": "helper-tool",
+  "version": "1.0.0",
+  "description": "A simple utility tool",
+  "main": "index.js",
+  "dependencies": {
+    "lodash": "^4.17.21"
+  }
+}"#,
+        )
+        .await?;
 
-async fn apply_filter_to_files(base_path: impl AsRef<OsStr>, filter: &Filter) -> Result<()> {
-    let is_bsd = is_bsd_sed().await;
+        // Execute the copy operation
+        copy_with_filters_impl(
+            &fs,
+            &src_path,
+            &dest_path,
+            &vec![
+                Filter::new(
+                    "**/*/package.json",
+                    r#""lodash":\s*"[^"]*""#,
+                    r#""lodash": "1.0.0""#,
+                ),
+                // Filter::new("package.json", "lodash", "nodash"),
+                Filter::new("**/*/file3.txt", "file", "chicken"),
+            ],
+        )
+        .await?;
 
-    // Find files matching the glob pattern
-    let find_output = Command::new("find")
-        .arg(base_path)
-        .arg("-name")
-        .arg(&filter.glob_pattern)
-        .arg("-type")
-        .arg("f")
-        .output()
-        .await
-        .context("Failed to execute find command")?;
+        // Verify that files were copied to the destination
+        // Check that the original files exist in source
+        assert!(fs
+            .read_to_string(format!("{}/file1.txt", src_path))
+            .await
+            .is_ok());
+        assert!(fs
+            .read_to_string(format!("{}/file2.txt", src_path))
+            .await
+            .is_ok());
+        assert!(fs
+            .read_to_string(format!("{}/subdir/file3.txt", src_path))
+            .await
+            .is_ok());
 
-    if !find_output.status.success() {
-        let stderr = String::from_utf8_lossy(&find_output.stderr);
-        bail!("find command failed: {}", stderr);
-    }
+        // Check that files were copied to destination
+        let copied_file1 = fs
+            .read_to_string(format!("{}/file1.txt", dest_path))
+            .await?;
+        let copied_file2 = fs
+            .read_to_string(format!("{}/file2.txt", dest_path))
+            .await?;
+        let copied_file3 = fs
+            .read_to_string(format!("{}/subdir/file3.txt", dest_path))
+            .await?;
+        let copied_package_json = fs
+            .read_to_string(format!(
+                "{}/tools/build/scripts/utils/helper-tool/package.json",
+                dest_path
+            ))
+            .await?;
+        let new_package_json = fs
+            .read_to_string(format!(
+                "{}/tools/build/scripts/utils/helper-tool/package.json",
+                dest_path
+            ))
+            .await?;
 
-    let files = String::from_utf8_lossy(&find_output.stdout);
-
-    // Apply sed replacement to each matching file
-    for file_path in files.lines().filter(|line| !line.is_empty()) {
-        let sed_cmd = format!(
-            "s|{}|{}|g",
-            escape_sed_pattern(&filter.search_pattern),
-            escape_sed_replacement(&filter.replacement)
+        assert_eq!(
+            new_package_json,
+            r#"{
+  "name": "helper-tool",
+  "version": "1.0.0",
+  "description": "A simple utility tool",
+  "main": "index.js",
+  "dependencies": {
+    "lodash": "1.0.0"
+  }
+}"#,
         );
 
-        println!("Running sed...");
-        println!("> {}", sed_cmd);
+        // Verify content matches
+        assert_eq!(copied_file1, "Content of file 1");
+        assert_eq!(copied_file2, "Content of file 2");
+        assert_eq!(copied_file3, "Content of chicken 3"); // substitution
+        assert!(copied_package_json.contains("helper-tool"));
+        assert!(copied_package_json.contains("lodash"));
 
-        let mut cmd = Command::new("sed");
-
-        // Check if we're on macOS (BSD sed) and add empty backup extension
-        if is_bsd {
-            // This is a quirk of BSD sed - we need to do the equivalent of:
-            // sed -i '' pattern filename
-            cmd.arg("-i").arg("");
-        } else {
-            cmd.arg("-i");
-        }
-
-        let sed_output = cmd
-            .arg(sed_cmd)
-            .arg(file_path)
-            .output()
-            .await
-            .context("Failed to execute sed command")?;
-
-        println!("{:?}", sed_output);
-
-        if !sed_output.status.success() {
-            let stderr = String::from_utf8_lossy(&sed_output.stderr);
-            bail!("sed command failed on {}: {}", file_path, stderr);
-        }
+        // Additional verification: check that the source files still exist (copy, not move)
+        let original_file1 = fs.read_to_string(format!("{}/file1.txt", src_path)).await?;
+        let original_file2 = fs.read_to_string(format!("{}/file2.txt", src_path)).await?;
+        let original_file3 = fs
+            .read_to_string(format!("{}/subdir/file3.txt", src_path))
+            .await?;
+        let original_package_json = fs
+            .read_to_string(format!(
+                "{}/tools/build/scripts/utils/helper-tool/package.json",
+                src_path
+            ))
+            .await?;
+        assert_eq!(original_file1, "Content of file 1");
+        assert_eq!(original_file2, "Content of file 2");
+        assert_eq!(original_file3, "Content of file 3");
+        assert!(original_package_json.contains("helper-tool"));
+        assert!(original_package_json.contains("lodash"));
+        Ok(())
     }
-
-    Ok(())
-}
-
-fn escape_sed_pattern(pattern: &str) -> String {
-    pattern.replace("|", "\\|") // Only escape the pipe delimiter
-}
-
-fn escape_sed_replacement(replacement: &str) -> String {
-    // Don't escape backslashes that are followed by digits (backreferences like \1, \2, etc.)
-    let mut result = replacement.to_string();
-
-    // First escape literal backslashes (but not backreferences)
-    // This is tricky - we need to escape \ that aren't part of \1, \2, etc.
-
-    // Simple approach: only escape the delimiter and ampersand
-    result = result.replace("|", "\\|"); // Escape pipe delimiter
-    result = result.replace("&", "\\&"); // Escape ampersand
-
-    result
 }
